Changing microcontroller and redoing WS2812 control
The QFN package required VIPPO to be easily soldered. I could
probably have done it without plating-over, but it still requires the
heat gun and it is currently broken.

The family was downgraded to an ST32F042, which is a little clunkier
than the STM32L052 family. It also doesn't have very accessible UARTs
when USB is being used, so I swapped out the level shifters to tri-state
ones and used the remaining GPIOs as enables.